\name{QCpipeline-package}
\alias{QCpipeline-package}
\alias{QCpipeline}
\docType{package}
\title{QC utility functions}

\description{
  This package contains functions for the QC pipeline.
}

\author{
  Stephanie Gogarten, Sarah Nelson, Jess Shen, Leila Zelnick
  
  Maintainer: Stephanie M. Gogarten \email{sdmorris@u.washington.edu}
}

\keyword{package}
