\name{readConfig}
\alias{readConfig}
\title{Read a configuration file}
\description{Read a configuration file}
\usage{
readConfig(file, ...)
}
\arguments{
  \item{file}{file where column 1 is parameter name
    and column 2 is value.}
  \item{...}{additional arguments to \code{read.table}}
}
\details{}
\value{
  Returns a named character vector of parameter values.
}
\author{Stephanie Gogarten}
\seealso{}
\examples{
}
\keyword{}
